EZP-22799: move Solr Storage initialization to dedicated bundle
