+++ eZ/Bundle/EzPublishSolrBundle/Command/SolrCreateIndexCommand.php
@@ -7,7 +7,7 @@
  * @version //autogentag//
  */
 
-namespace eZ\Bundle\EzPublishCoreBundle\Command;
+namespace eZ\Bundle\EzPublishSolrBundle\Command;
 
 use Symfony\Bundle\FrameworkBundle\Command\ContainerAwareCommand;
 use Symfony\Component\Console\Input\InputInterface;
