+++ eZ/Bundle/EzPublishSolrBundle/DependencyInjection/Compiler/AggregateSortClauseVisitorPass.php
@@ -7,7 +7,7 @@
  * @version //autogentag//
  */
 
-namespace eZ\Bundle\EzPublishCoreBundle\DependencyInjection\Compiler\Storage\Solr;
+namespace eZ\Bundle\EzPublishSolrBundle\DependencyInjection\Compiler;
 
 use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
