+++ eZ/Bundle/EzPublishSolrBundle/DependencyInjection/EzPublishSolrExtension.php
@@ -1,0 +1,38 @@
+<?php
+/**
+ * File containing the EzPublishSolrExtension class.
+ *
+ * @copyright Copyright (C) 1999-2014 eZ Systems AS. All rights reserved.
+ * @license http://www.gnu.org/licenses/gpl-2.0.txt GNU General Public License v2
+ * @version //autogentag//
+ */
+
+namespace eZ\Bundle\EzPublishSolrBundle\DependencyInjection;
+
+use Symfony\Component\HttpKernel\DependencyInjection\Extension;
+use Symfony\Component\DependencyInjection\ContainerBuilder;
+use Symfony\Component\DependencyInjection\Loader\YamlFileLoader;
+use Symfony\Component\Config\FileLocator;
+
+class EzPublishSolrExtension extends Extension
+{
+    /**
+     * Loads a specific configuration.
+     *
+     * @param array $configs An array of configuration values
+     * @param ContainerBuilder $container A ContainerBuilder instance
+     *
+     * @throws \InvalidArgumentException When provided tag is not defined in this extension
+     *
+     * @api
+     */
+    public function load( array $configs, ContainerBuilder $container )
+    {
+        $loader = new YamlFileLoader(
+            $container,
+            new FileLocator( __DIR__ . '/../Resources/config' )
+        );
+        $loader->load( 'indexable_fieldtypes.yml' );
+        $loader->load( "storage_engines/legacy_solr.yml" );
+    }
+}
