+++ eZ/Bundle/EzPublishSolrBundle/Resources/config/indexable_fieldtypes.yml
