+++ eZ/Bundle/EzPublishSolrBundle/Resources/config/storage_engines/legacy_solr.yml
