+++ eZ/Bundle/EzPublishSolrBundle/Resources/config/storage_engines/solr/criterion_visitors.yml
