+++ eZ/Bundle/EzPublishSolrBundle/Resources/config/storage_engines/solr/facet_builder_visitors.yml
