+++ eZ/Bundle/EzPublishSolrBundle/Resources/config/storage_engines/solr/field_value_mappers.yml
