+++ eZ/Bundle/EzPublishSolrBundle/Resources/config/storage_engines/solr/services.yml
