+++ eZ/Bundle/EzPublishSolrBundle/Resources/config/storage_engines/solr/slots.yml
