+++ eZ/Bundle/EzPublishSolrBundle/Resources/config/storage_engines/solr/sort_clause_visitors.yml
