+++ eZ/Bundle/EzPublishSolrBundle/Tests/DependencyInjection/Compiler/AggregateSortClauseVisitorPassTest.php
@@ -7,9 +7,9 @@
  * @version //autogentag//
  */
 
-namespace eZ\Bundle\EzPublishCoreBundle\Tests\DependencyInjection\Compiler\Storage\Solr;
+namespace eZ\Bundle\EzPublishSolrBundle\Tests\DependencyInjection\Compiler;
 
-use eZ\Bundle\EzPublishCoreBundle\DependencyInjection\Compiler\Storage\Solr\AggregateSortClauseVisitorPass;
+use eZ\Bundle\EzPublishSolrBundle\DependencyInjection\Compiler\AggregateSortClauseVisitorPass;
 use Matthias\SymfonyDependencyInjectionTest\PhpUnit\AbstractCompilerPassTest;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Definition;
