+++ eZ/Bundle/EzPublishSolrBundle/Tests/DependencyInjection/Compiler/FieldTypeRegistryPassTest.php
@@ -1,26 +1,26 @@
 <?php
 /**
- * File containing the SignalSlotPassTest class.
+ * File containing the FieldTypeRegistryPassTest class.
  *
  * @copyright Copyright (C) 1999-2014 eZ Systems AS. All rights reserved.
  * @license http://www.gnu.org/licenses/gpl-2.0.txt GNU General Public License v2
  * @version //autogentag//
  */
 
-namespace eZ\Bundle\EzPublishCoreBundle\Tests\DependencyInjection\Compiler\Storage\Solr;
+namespace eZ\Bundle\EzPublishSolrBundle\Tests\DependencyInjection\Compiler;
 
-use eZ\Bundle\EzPublishCoreBundle\DependencyInjection\Compiler\Storage\Solr\SignalSlotPass;
+use eZ\Bundle\EzPublishSolrBundle\DependencyInjection\Compiler\FieldRegistryPass;
 use Matthias\SymfonyDependencyInjectionTest\PhpUnit\AbstractCompilerPassTest;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Definition;
 use Symfony\Component\DependencyInjection\Reference;
 
-class SignalSlotPassTest extends AbstractCompilerPassTest
+class FieldTypeRegistryPassTest extends AbstractCompilerPassTest
 {
     protected function setUp()
     {
         parent::setUp();
-        $this->setDefinition( 'ezpublish.signalslot.signal_dispatcher', new Definition() );
+        $this->setDefinition( 'ezpublish.persistence.solr.search.field_registry', new Definition() );
     }
 
     /**
@@ -31,43 +31,43 @@ class SignalSlotPassTest extends AbstractCompilerPassTest
      */
     protected function registerCompilerPass( ContainerBuilder $container )
     {
-        $container->addCompilerPass( new SignalSlotPass() );
+        $container->addCompilerPass( new FieldRegistryPass() );
     }
 
-    public function testAttachSignal()
+    public function testRegisterFieldType()
     {
-        $signal = 'signal_identifier';
+        $fieldTypeIdentifier = 'field_type_identifier';
         $serviceId = 'service_id';
         $def = new Definition();
-        $def->addTag( 'ezpublish.persistence.solr.slot', array( 'signal' => $signal ) );
+        $def->addTag( 'ezpublish.fieldType.indexable', array( 'alias' => $fieldTypeIdentifier ) );
         $this->setDefinition( $serviceId, $def );
 
         $this->compile();
 
         $this->assertContainerBuilderHasServiceDefinitionWithMethodCall(
-            'ezpublish.signalslot.signal_dispatcher',
-            'attach',
-            array( $signal, new Reference( $serviceId ) )
+            'ezpublish.persistence.solr.search.field_registry',
+            'registerType',
+            array( $fieldTypeIdentifier, new Reference( $serviceId ) )
         );
     }
 
     /**
      * @expectedException \LogicException
      */
-    public function testAttachSignalNoAlias()
+    public function testRegisterFieldTypeNoAlias()
     {
-        $signal = 'signal_identifier';
+        $fieldTypeIdentifier = 'field_type_identifier';
         $serviceId = 'service_id';
         $def = new Definition();
-        $def->addTag( 'ezpublish.persistence.solr.slot' );
+        $def->addTag( 'ezpublish.fieldType.indexable' );
         $this->setDefinition( $serviceId, $def );
 
         $this->compile();
 
         $this->assertContainerBuilderHasServiceDefinitionWithMethodCall(
-            'ezpublish.signalslot.signal_dispatcher',
-            'attach',
-            array( $signal, new Reference( $serviceId ) )
+            'ezpublish.persistence.solr.search.field_registry',
+            'registerType',
+            array( $fieldTypeIdentifier, new Reference( $serviceId ) )
         );
     }
 }
