+++ eZ/Bundle/EzPublishSolrBundle/Tests/DependencyInjection/Compiler/SignalSlotPassTest.php
@@ -1,26 +1,26 @@
 <?php
 /**
- * File containing the FieldTypeRegistryPassTest class.
+ * File containing the SignalSlotPassTest class.
  *
  * @copyright Copyright (C) 1999-2014 eZ Systems AS. All rights reserved.
  * @license http://www.gnu.org/licenses/gpl-2.0.txt GNU General Public License v2
  * @version //autogentag//
  */
 
-namespace eZ\Bundle\EzPublishCoreBundle\Tests\DependencyInjection\Compiler\Storage\Solr;
+namespace eZ\Bundle\EzPublishSolrBundle\Tests\DependencyInjection\Compiler;
 
-use eZ\Bundle\EzPublishCoreBundle\DependencyInjection\Compiler\Storage\Solr\FieldRegistryPass;
+use eZ\Bundle\EzPublishSolrBundle\DependencyInjection\Compiler\SignalSlotPass;
 use Matthias\SymfonyDependencyInjectionTest\PhpUnit\AbstractCompilerPassTest;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Definition;
 use Symfony\Component\DependencyInjection\Reference;
 
-class FieldTypeRegistryPassTest extends AbstractCompilerPassTest
+class SignalSlotPassTest extends AbstractCompilerPassTest
 {
     protected function setUp()
     {
         parent::setUp();
-        $this->setDefinition( 'ezpublish.persistence.solr.search.field_registry', new Definition() );
+        $this->setDefinition( 'ezpublish.signalslot.signal_dispatcher', new Definition() );
     }
 
     /**
@@ -31,43 +31,43 @@ class FieldTypeRegistryPassTest extends AbstractCompilerPassTest
      */
     protected function registerCompilerPass( ContainerBuilder $container )
     {
-        $container->addCompilerPass( new FieldRegistryPass() );
+        $container->addCompilerPass( new SignalSlotPass() );
     }
 
-    public function testRegisterFieldType()
+    public function testAttachSignal()
     {
-        $fieldTypeIdentifier = 'field_type_identifier';
+        $signal = 'signal_identifier';
         $serviceId = 'service_id';
         $def = new Definition();
-        $def->addTag( 'ezpublish.fieldType.indexable', array( 'alias' => $fieldTypeIdentifier ) );
+        $def->addTag( 'ezpublish.persistence.solr.slot', array( 'signal' => $signal ) );
         $this->setDefinition( $serviceId, $def );
 
         $this->compile();
 
         $this->assertContainerBuilderHasServiceDefinitionWithMethodCall(
-            'ezpublish.persistence.solr.search.field_registry',
-            'registerType',
-            array( $fieldTypeIdentifier, new Reference( $serviceId ) )
+            'ezpublish.signalslot.signal_dispatcher',
+            'attach',
+            array( $signal, new Reference( $serviceId ) )
         );
     }
 
     /**
      * @expectedException \LogicException
      */
-    public function testRegisterFieldTypeNoAlias()
+    public function testAttachSignalNoAlias()
     {
-        $fieldTypeIdentifier = 'field_type_identifier';
+        $signal = 'signal_identifier';
         $serviceId = 'service_id';
         $def = new Definition();
-        $def->addTag( 'ezpublish.fieldType.indexable' );
+        $def->addTag( 'ezpublish.persistence.solr.slot' );
         $this->setDefinition( $serviceId, $def );
 
         $this->compile();
 
         $this->assertContainerBuilderHasServiceDefinitionWithMethodCall(
-            'ezpublish.persistence.solr.search.field_registry',
-            'registerType',
-            array( $fieldTypeIdentifier, new Reference( $serviceId ) )
+            'ezpublish.signalslot.signal_dispatcher',
+            'attach',
+            array( $signal, new Reference( $serviceId ) )
         );
     }
 }
